Move McpTransportContext to common package, make impl immutable
