+++ mcp/src/main/java/io/modelcontextprotocol/common/DefaultMcpTransportContext.java
@@ -2,7 +2,7 @@
  * Copyright 2024-2025 the original author or authors.
  */
 
-package io.modelcontextprotocol.server;
+package io.modelcontextprotocol.common;
 
 import java.util.Map;
