+++ mcp/src/main/java/io/modelcontextprotocol/server/McpAsyncServerExchange.java
@@ -4,6 +4,7 @@
 
 package io.modelcontextprotocol.server;
 
+import io.modelcontextprotocol.common.McpTransportContext;
 import java.util.ArrayList;
 import java.util.Collections;
