+++ mcp/src/main/java/io/modelcontextprotocol/server/McpStatelessServerHandler.java
@@ -4,6 +4,7 @@
 
 package io.modelcontextprotocol.server;
 
+import io.modelcontextprotocol.common.McpTransportContext;
 import io.modelcontextprotocol.spec.McpSchema;
 import reactor.core.publisher.Mono;
