+++ mcp/src/main/java/io/modelcontextprotocol/server/McpTransportContextExtractor.java
@@ -4,6 +4,8 @@
 
 package io.modelcontextprotocol.server;
 
+import io.modelcontextprotocol.common.McpTransportContext;
+
 /**
  * The contract for extracting metadata from a generic transport request of type
  * {@link T}.
